+++ src/main/scala/longevity/exceptions/subdomain/DerivedHasNoPolyException.scala
@@ -2,8 +2,8 @@ package longevity.exceptions.subdomain
 
 /** thrown on attempt to construct a
  * [[longevity.subdomain.Subdomain subdomain]] with a
- * [[longevity.subdomain.embeddable.DerivedType derived type]] that does not
- * have a corresponding [[longevity.subdomain.embeddable.PolyType poly type]],
+ * [[longevity.subdomain.embeddable.DerivedEType derived type]] that does not
+ * have a corresponding [[longevity.subdomain.embeddable.PolyEType poly type]],
  * or a [[longevity.subdomain.ptype.DerivedPType derived persistent type]] that
  * does not have a corresponding [[longevity.subdomain.ptype.PolyPType
  * poly persistent type]]
@@ -11,4 +11,4 @@ package longevity.exceptions.subdomain
 class DerivedHasNoPolyException(typeName: String, isPType: Boolean)
 extends SubdomainException(
   if (isPType) "DerivedPType $typeName does not have a corresponding PolyPType"
-  else "DerivedType $typeName does not have a corresponding PolyPype")
+  else "DerivedEType $typeName does not have a corresponding PolyPype")