+++ src/test/scala/longevity/integration/poly/CassandraPolyReposSpec.scala
@@ -6,7 +6,7 @@ import longevity.context.Cassandra
 import longevity.context.LongevityContext
 import longevity.integration.subdomain.derivedEntities
 
-/** tests for cassandra repos that share tables in the presence of [[PolyType]] */
+/** tests for cassandra repos that share tables in the presence of [[PolyEType]] */
 class CassandraPolyReposSpec extends PolyReposSpec(
   new LongevityContext(
     derivedEntities.subdomain,