+++ src/test/scala/longevity/integration/poly/MongoPolyReposSpec.scala
@@ -6,7 +6,7 @@ import longevity.context.LongevityContext
 import longevity.context.Mongo
 import longevity.integration.subdomain.derivedEntities
 
-/** tests for mongo repos that share tables in the presence of [[PolyType]] */
+/** tests for mongo repos that share tables in the presence of [[PolyEType]] */
 class MongoPolyReposSpec extends PolyReposSpec(
   new LongevityContext(
     derivedEntities.subdomain,