+++ src/test/scala/longevity/integration/subdomain/controlledVocab/package.scala
@@ -2,9 +2,9 @@ package longevity.integration.subdomain
 
 import longevity.TestLongevityConfigs
 import longevity.subdomain.Subdomain
-import longevity.subdomain.embeddable.DerivedType
+import longevity.subdomain.embeddable.DerivedEType
 import longevity.subdomain.embeddable.ETypePool
-import longevity.subdomain.embeddable.PolyType
+import longevity.subdomain.embeddable.PolyEType
 import longevity.subdomain.ptype.PTypePool
 
 /** covers a controlled vocab created with a poly type and multiple derived case objects */
@@ -14,9 +14,9 @@ package object controlledVocab {
     "Controlled Vocabulary",
     PTypePool(WithControlledVocab),
     ETypePool(
-      PolyType[ControlledVocab],
-      DerivedType[VocabTerm1, ControlledVocab],
-      DerivedType[VocabTerm2.type, ControlledVocab]))
+      PolyEType[ControlledVocab],
+      DerivedEType[VocabTerm1, ControlledVocab],
+      DerivedEType[VocabTerm2.type, ControlledVocab]))
 
   val contexts = TestLongevityConfigs.sparseContextMatrix(subdomain)