+++ src/test/scala/longevity/integration/subdomain/derivedEntities/package.scala
@@ -1,9 +1,9 @@
 package longevity.integration.subdomain
 
 import longevity.TestLongevityConfigs
-import longevity.subdomain.embeddable.DerivedType
+import longevity.subdomain.embeddable.DerivedEType
 import longevity.subdomain.embeddable.ETypePool
-import longevity.subdomain.embeddable.PolyType
+import longevity.subdomain.embeddable.PolyEType
 import longevity.subdomain.Subdomain
 import longevity.subdomain.ptype.PTypePool
 
@@ -14,9 +14,9 @@ package object derivedEntities {
     "Derived Entities",
     PTypePool(PolyRoot, FirstDerivedRoot, SecondDerivedRoot),
     ETypePool(
-      PolyType[PolyEntity],
-      DerivedType[FirstDerivedEntity, PolyEntity],
-      DerivedType[SecondDerivedEntity, PolyEntity]))
+      PolyEType[PolyEntity],
+      DerivedEType[FirstDerivedEntity, PolyEntity],
+      DerivedEType[SecondDerivedEntity, PolyEntity]))
 
   val contexts = TestLongevityConfigs.sparseContextMatrix(subdomain)